Split out common base for integration tests
